@@ -16,25 +16,22 @@
 package meta
 
 import (
-	"io/ioutil"
+	"io"
 	"log"
 	"time"
 
 	"github.com/segmentio/analytics-go/v3"
 )
 
-var (
-	SEGMENT_CLIENT analytics.Client
-)
+var SEGMENT_CLIENT analytics.Client
 
 func init() {
-
 	cf := analytics.Config{
 		Interval:  15 * time.Second,
 		BatchSize: 10,
 		Endpoint:  "https://e1.zinclabs.dev",
 		Verbose:   false,
-		Logger:    analytics.StdLogger(log.New(ioutil.Discard, "marker ", log.LstdFlags)), // discard any logs
+		Logger:    analytics.StdLogger(log.New(io.Discard, "marker ", log.LstdFlags)), // discard any logs
 	}
 
 	SEGMENT_CLIENT, _ = analytics.NewWithConfig("", cf)
