@@ -63,7 +63,7 @@ func TestLoadDict(t *testing.T) {
 	}
 
 	t.Run("prepare dict", func(t *testing.T) {
-		_ = os.Mkdir("data", 0755)
+		_ = os.Mkdir("data", 0o755)
 		config.Global.Plugin.GSE.DictPath = "./data"
 		err := writeFile("./data/user.txt", "你若安好便是晴天 100 n\n")
 		assert.NoError(t, err)
@@ -189,7 +189,7 @@ func TestNewGseStopTokenFilter(t *testing.T) {
 }
 
 func writeFile(path string, content string) error {
-	f, err := os.OpenFile(path, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0644)
+	f, err := os.OpenFile(path, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0o644)
 	if err != nil {
 		return err
 	}
