@@ -68,8 +68,8 @@ type cluster struct {
 }
 
 type shard struct {
-	// control gorutine number for read
-	GorutineNum int `env:"ZINC_SHARD_GORUTINE_NUM,default=3"`
+	// control goroutine number for read
+	GoroutineNum int `env:"ZINC_SHARD_GOROUTINE_NUM,default=3"`
 	// DefaultNum is the default number of shards.
 	Num int64 `env:"ZINC_SHARD_NUM,default=3"`
 	// MaxSize is the maximum size limit for one shard, or will create a new shard.
@@ -116,7 +116,7 @@ func init() {
 
 	// check data path
 	testPath := path.Join(Global.DataPath, "_test_")
-	if err := os.MkdirAll(testPath, 0755); err != nil {
+	if err := os.MkdirAll(testPath, 0o755); err != nil {
 		log.Fatal().Err(err).Msg("ZINC_DATA_PATH is not writable")
 	}
 	if err := os.Remove(testPath); err != nil {
