@@ -167,7 +167,7 @@ func (t *IndexList) Close() error {
 	t.lock.Lock()
 	defer t.lock.Unlock()
 	eg := errgroup.Group{}
-	eg.SetLimit(config.Global.Shard.GorutineNum)
+	eg.SetLimit(config.Global.Shard.GoroutineNum)
 	for _, index := range t.Indexes {
 		index := index
 		eg.Go(func() error {
