@@ -259,8 +259,8 @@ func TestIndex_Search(t *testing.T) {
 		})
 
 		for _, d := range prepareData {
-			rand.Seed(time.Now().UnixNano())
-			docId := rand.Intn(1000)
+			r := rand.New(rand.NewSource(time.Now().UnixNano()))
+			docId := r.Intn(1000)
 			err := index.CreateDocument(strconv.Itoa(docId), d, false)
 			assert.NoError(t, err)
 		}
