@@ -46,10 +46,10 @@ func openbboltDB(dbpath string, readOnly bool) (*bbolt.DB, error) {
 		NoGrowSync:   false,
 		FreelistType: bbolt.FreelistArrayType,
 	}
-	if err := os.MkdirAll(path.Dir(dbpath), 0755); err != nil {
+	if err := os.MkdirAll(path.Dir(dbpath), 0o755); err != nil {
 		return nil, err
 	}
-	return bbolt.Open(dbpath, 0666, opt)
+	return bbolt.Open(dbpath, 0o666, opt)
 }
 
 func (t *boltStorage) List(prefix string, _, _ int) ([][]byte, error) {
