@@ -21,10 +21,10 @@ import (
 
 	"github.com/blugelabs/bluge"
 	"github.com/blugelabs/bluge/analysis"
-
 	"github.com/zincsearch/zincsearch/pkg/errors"
 	"github.com/zincsearch/zincsearch/pkg/meta"
 	zincanalysis "github.com/zincsearch/zincsearch/pkg/uquery/analysis"
+	zincanalyzer "github.com/zincsearch/zincsearch/pkg/uquery/analysis/analyzer"
 	"github.com/zincsearch/zincsearch/pkg/zutils"
 )
 
@@ -36,6 +36,7 @@ func MatchQuery(query map[string]interface{}, mappings *meta.Mappings, analyzers
 	field := ""
 	value := new(meta.MatchQuery)
 	value.Boost = -1.0
+	var minimumShouldMatch interface{}
 	for k, v := range query {
 		field = k
 		switch v := v.(type) {
@@ -57,6 +58,8 @@ func MatchQuery(query map[string]interface{}, mappings *meta.Mappings, analyzers
 					value.PrefixLength, _ = zutils.ToFloat64(v)
 				case "boost":
 					value.Boost, _ = zutils.ToFloat64(v)
+				case "minimum_should_match":
+					minimumShouldMatch = v
 				default:
 					// return nil, errors.New(errors.ErrorTypeParsingException, fmt.Sprintf("[match] unknown field [%s]", k))
 				}
@@ -81,6 +84,11 @@ func MatchQuery(query map[string]interface{}, mappings *meta.Mappings, analyzers
 		if zer == nil && indexZer != nil {
 			zer = indexZer
 		}
+	}
+
+	// only "OR" supports minimum should match
+	if minimumShouldMatch != nil && (value.Operator == "" || strings.ToUpper(value.Operator) == "OR") {
+		return genQueryWithMinimumShouldMatch(zer, field, value, minimumShouldMatch)
 	}
 
 	subq := bluge.NewMatchQuery(value.Query).SetField(field)
@@ -114,4 +122,57 @@ func MatchQuery(query map[string]interface{}, mappings *meta.Mappings, analyzers
 	}
 
 	return subq, nil
+}
+
+func genQueryWithMinimumShouldMatch(ana *analysis.Analyzer, field string, value *meta.MatchQuery, minimumShouldMatch interface{}) (bluge.Query, error) {
+	if ana == nil {
+		ana, _ = zincanalyzer.NewStandardAnalyzer(nil)
+	}
+
+	var fuzziness int
+	if value.Fuzziness != nil {
+		if value.Fuzziness != nil {
+			v := ParseFuzziness(value.Fuzziness, value.Query, ana)
+			if v > 0 {
+				fuzziness = v
+			}
+		}
+	}
+
+	var boost float64 = 1
+	if value.Boost >= 0 {
+		boost = value.Boost
+	}
+
+	tokens := ana.Analyze([]byte(value.Query))
+	if len(tokens) > 0 {
+		tqs := make([]bluge.Query, len(tokens))
+		if fuzziness != 0 {
+			for i, token := range tokens {
+				query := bluge.NewFuzzyQuery(string(token.Term))
+				query.SetFuzziness(fuzziness)
+				query.SetPrefix(int(value.PrefixLength))
+				query.SetField(field)
+				query.SetBoost(boost)
+				tqs[i] = query
+			}
+		} else {
+			for i, token := range tokens {
+				tq := bluge.NewTermQuery(string(token.Term))
+				tq.SetField(field)
+				tq.SetBoost(boost)
+				tqs[i] = tq
+			}
+		}
+		minValue, err := zutils.CalculateMin(len(tokens), minimumShouldMatch)
+		if err != nil {
+			return nil, err
+		}
+		booleanQuery := bluge.NewBooleanQuery()
+		booleanQuery.AddShould(tqs...)
+		booleanQuery.SetMinShould(minValue)
+		booleanQuery.SetBoost(boost)
+		return booleanQuery, nil
+	}
+	return bluge.NewMatchNoneQuery(), nil
 }
