@@ -35,13 +35,14 @@ func Open(indexName string) (*Log, error) {
 	l := new(Log)
 	l.name = indexName
 	opt := &wal.Options{
-		NoSync:           true,     // Fsync after every write
-		SegmentSize:      16777216, // 16 MB log segment files.
-		SegmentCacheSize: 2,        // Number of cached in-memory segments
-		NoCopy:           true,     // Make a new copy of data for every Read call.
-		DirPerms:         0750,     // Permissions for the created directories
-		FilePerms:        0640,     // Permissions for the created data files
-		FillID:           true,     // Allow writes with a zero ID
+		NoSync:               true,     // Fsync after every write
+		SegmentSize:          16777216, // 16 MB log segment files.
+		SegmentCacheSize:     2,        // Number of cached in-memory segments
+		NoCopy:               true,     // Make a new copy of data for every Read call.
+		DirPerms:             0o750,     // Permissions for the created directories
+		FilePerms:            0o640,     // Permissions for the created data files
+		FillID:               true,     // Allow writes with a zero ID
+		RecoverCorruptedTail: true,     // Enables corruption recovery
 	}
 	l.log, err = wal.Open(path.Join(config.Global.DataPath, indexName, "wal"), opt)
 	if err != nil {
